requirements.txt
slack_sdk>=3.15.0
Flask>=2.0.0
python-dotenv>=0.19.0
gunicorn==20.1.0
requests==2.28.2
werkzeug>=2.0.0
